Update abc.R
Dear Frédéric Bertrand and Myriam Maumy-Bertrand,

I hope this message finds you well. 

I'm using networkABC in some analysis and i wondered why the matrix should be limited to 30 rows and one column. I know it's the matrix used in the example but for real data, I can't find a reason for this formatting.